feat: Vector upstream 0.32.2

+++ src/transforms/mezmo_tag_cardinality_limit/config.rs
@@ -14,7 +14,7 @@ use super::TagCardinalityLimit;
 pub struct TagCardinalityLimitConfig {
     /// How many distinct values to accept for any given key.
     #[serde(default = "default_value_limit")]
-    pub value_limit: u32,
+    pub value_limit: usize,
 
     #[configurable(derived)]
     #[serde(default = "default_limit_exceeded_action")]
@@ -91,7 +91,7 @@ const fn default_limit_exceeded_action() -> LimitExceededAction {
     LimitExceededAction::DropTag
 }
 
-const fn default_value_limit() -> u32 {
+const fn default_value_limit() -> usize {
     500
 }
